requirements.txt
termcolor==1.1.0
numpy==1.19.5
matplotlib==3.4.2
pyfiglet==0.8.post1
colorama==0.4.4
Pillow==8.4.0
torch==1.10.0
torchvision==0.11.1
